+++ apps/pack.py
@@ -23,6 +23,9 @@ def pack():
         return render_template('page-login.html')
     elif 'lock_stat' in session:
         return redirect(url_for('lock'))
+    nodes = Node.query.filter().all()
+    if len(nodes) == 0:
+        return redirect(url_for('sys_set'))
     if os.path.exists(TEMP_FOLDER):
         shutil.rmtree(TEMP_FOLDER)
     os.mkdir(TEMP_FOLDER)
